core/state_manager/state_manager_interface.py
import grpc
import core.proto.state_manager_pb2 as state_manager_pb2
import core.proto.state_manager_pb2_grpc as state_manager_pb2_grpc

class StateManagerService(state_manager_pb2_grpc.StateManagerServicer):
    """
    Implementation of the StateManager gRPC service.
    """
    def __init__(self):
        # TODO: Initialize Git and Redis clients
        pass

    def GetDefinitionFileContent(self, request, context):
        # TODO: Implement logic to get file content from Git
        pass

    def SetDefinitionFileContent(self, request, context):
        # TODO: Implement logic to set file content and commit to Git
        pass

    def DeleteDefinitionFile(self, request, context):
        # TODO: Implement logic to delete file and commit to Git
        pass

    def ListDefinitionDirectory(self, request, context):
        # TODO: Implement logic to list directory contents from Git
        pass

    def GetRuntimeValue(self, request, context):
        # TODO: Implement logic to get value from Redis
        pass

    def SetRuntimeValue(self, request, context):
        # TODO: Implement logic to set value in Redis
        pass

    def DeleteRuntimeValue(self, request, context):
        # TODO: Implement logic to delete value from Redis
        pass

import time
from concurrent import futures

def serve():
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    state_manager_pb2_grpc.add_StateManagerServicer_to_server(
        StateManagerService(), server)
    # TODO: Make port configurable
    server.add_insecure_port('[::]:50051')
    server.start()
    print("StateManager gRPC server started on port 50051")
    try:
        while True:
            time.sleep(86400) # Serve for 1 day
    except KeyboardInterrupt:
        server.stop(0)

if __name__ == '__main__':
    serve()
